Move files service to shared
